create lobby reducer, move onlinePlayers into it
