+++ packages/frontend/store/lobby/reducer.ts
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux';
+
+import onlinePlayers from './onlinePlayers/reducers';
+
+const lobbyReducer = combineReducers({
+  onlinePlayers,
+});
+
+export default lobbyReducer;
